dummysigner: stakeholder_client
